backend/requirements.txt
django
djangorestframework
black
django-cors-headers
pypdf2